pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "ipykernel>=7.1.0",
    "langchain>=1.2.8",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.2",
    "pinecone>=8.0.0",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "reportlab>=4.4.9",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.2",
    "streamlit>=1.53.1",
    "uvicorn>=0.40.0",
]
